Remove HarrisGrammar, move its contents to comp.y
HarrisGrammar.y was only there so that we would have the full C grammar
and token names for everybody to use.  Now that the grammar seems to be
reformatted and all the tokens renamed as desired, I've just taken that
grammar and moved it into our actual Yacc file, so HarrisGrammar.y is
obsolete.

Typing 'make' now seems to compile and create the executable fine with
comp.y.

Note: We do get a warning that says there is 1 shift/reduce conflict.
I'm not sure whether or not this is expected.  We'll have to look into
it.