Print reductions to LIST_FILE from parser
I've made a const string called LIST_FILE containing the file name to
output reductions to.  It is set to "list_file"

I've made a new function called reductionOut that takes a string
(presumably a string showing the current reduction) and outputs it both
to LIST_FILE and optionally to parser debugging.